Correct error on md-empty-state on Professor View
the empty-state component was duplicated because v-for takes precedence over v-if so I re-arranged that section of the page.

I also made my placeholder data for ratings look closer to the real thing so I could put together the design of the ProfessorReview component.

I removed 'avgRatings' because I won't have such a data. It will be computed on the front-end.