Remove state interactions from the NavbarSearchForm Component
Since I've added a 'loading spinner' with the search bar to make it easier to understand what's happening, I needed to 'pull' the data fetching from this component to the one that above it (the Navbar) so that I can decide what to render and when.

So instead I'm going to be passing down the search data as props.