Add a loading spinner on top of search bar in Navbar Component
Since data and such are going to be pulled from the api I need to have a loading spinner to indicate to users what is happening.

Because of that, I needed to move up the data fetching from the NavbarSearchForm component to this one so I can control what to render and when.

PS: I've added the documentation as a bonus because it was hurting my eyes to figure what was happening and procrastinating to 'document everything later' is a trap.