Partially complete the search form for professors in QuickSearch component
I laid the groundwork and I did a few things like setting up the autocomplete and adding comments to component.

I also changed the way the autocomplete worked where you had to use your keyboard arrows to navigate suggestions. I realized that most people that will be using this website are probably not going to be tech-savy to think about using it and instead of having added complexity to try to handle the hovers + arrow nav I decided to just ditch the arrow nav.