Update auth@updateAccessToken migrations to clear expires
The way I had this previous set was that even I wanted to clear the state and I set expire to null, it would still have a value of Date.now().

So now I check that expire is truthy before calculating the expire time in miliseconds or else set to null.