Update user middleware to pass down params too
When a user was being redirected off for trying to access a page they aren't allowed to see. The page they were coming from was kept, but the params were not.

So what would happen is after logging in, the app would try to redirect back to the page it came from but without the params a 404 error would be shown instead of the originally seeked page.

Therefore, I decided to add the params as query arguments as well so they can be kept and used for correct redirecting.