Add url params to the redirect after login in SignIn page
I've added an URL's params to the query whenever the `userOnly` middleware sends a user to this page to make sure that I'm able to send them back to the correct page after they login.

For example, if I was coming from a prof's page, after success login I would be sent to a 404 not found page because I would try to redirect to a page without the slug. So now this type of issue has been fixed.