Add auth method for restricted pages in store/api
I created this method to extract the access token from the store's state and return the approriate headers because there are a lot of pages that are restricted to users only and it would be tedious having to keep track of all that. It's not a DRY way to code so I created this method.