Re-design ProfessorReviewController@create
I can't use the request->only() method because I need to pass the user id to the professor review entry so I modified the method.

I also added verification to update() and delete() to verify that a user has right to modify an entry before moving forward.

I also got rid of the LaravelController because it isn't needed since there is no need for a fetch() method with advanced params.