Simplify DetailsCard component
I was about to start working on getting real data showing and I didn't want to create 5 computed props to calculate the average considering they all look so similar so I created a method to figure out the average instead.

I then made an computed prop of the averages so that I can use a v-for to render them all instead of copy/pasting 5 indentical lines.