Create the Professor/ReviewFeedback model
The reason why I made it restricted to professor instead of trying to make a global ReviewFeedback model is because I would have to have a column that says 'profile' that would either be 'professor' or 'school' and I feel like that's not the cleanest solution. I would rather have 2 different tables than one table to do 2 things.