Shift responsability of vote handling from Professor & School Review to ReviewFeedback component
It's kind of a waste if I have a ReviewFeedback component and it can't handle to get and set the votes on it's own and puts a additional charge on it's parent component.

I decided to make that shift because I noticed that I had a duplicating code on ProfessorReview and SchoolReview that I could consolidate and DRY out, so I did.